added validation for order and quantity

+++ Data/Repo/ProductRepository.cpp
@@ -119,7 +119,7 @@ namespace productrepo {
         return false;
     }
 
-    bool ProductRepository::validate_stock(int id, int amount)
+    bool ProductRepository::validate_stock(int id, float amount)
     {
         if (amount > 0)
         {
@@ -127,7 +127,7 @@ namespace productrepo {
             {
                 if ((*it)->get_id() == std::to_string(id))
                 {
-                    if (float(amount) <= (*it)->get_quantity())
+                    if (amount <= (*it)->get_quantity())
                     {
                         return true;
                     }
