+++ Service/Controller/ProductController.cpp
@@ -47,7 +47,7 @@ bool ProductController::validate_id (int id)
     return productRepository.validate_id(id);
 }
 
-bool ProductController::validate_stock(int id, int amount)
+bool ProductController::validate_stock(int id, float amount)
 {
     return productRepository.validate_stock(id,amount);
 }
